Do not reset canTrigger so easily
I used this sketch in a Kong Wobbler and found that it frequently triggered
dispensing 2 or 3 times per "wobble". The current code waits for it to become
stable, but I think that it can read below the threshold at the edges of
a wobble, and this only has to happen once to reset the `canTrigger` flag.

With these changes, it must read as stable (below the threshold) for a number
of iterations (I've been happy with 5).