Create PetTutor_Remote_motionV1
This file is a modified version of the button-triggered remote to control the PTFeeder.  It uses the gyroscope feature of the Arduino Nano 33 BLE.  Timing has been adjusted a bit but there are still issues with initial connection and sometimes the connection drops requiring a reset via software (I use PunchThrough LightBlue app) or power-cycling the feeder.